Fix MISRAC for the folder include
 - MISRA.ETYPE.INAPPR.CAST.2012
  - The value of an expression should not be cast to
    an inappropriate essential type
  - Fixed by typecasting the parameter cond to int

 - MISRA.ASSIGN.SUBEXPR.2012
  - The result of an assignment operator should not be used
  - Fixed by Splitting the expression

 - MISRA.RETURN.NOT_LAST
  - A function should have a single point of exit at the end
  - Fixed by a new variable for return values/expressions

 - MISRA.BUILTIN_NUMERIC
  - Builtin numeric type should not be used
  - typedefs that indicate size and signedness should be used
    in place of the basic numerical types

 - MISRA.RETURN.NOT_LAST
  - A function should have a single point of exit at the end
  - Fixed by a new variable for return values/expressions

- MISRA.ETYPE.COMP.ASSIGN.2012
  - The value of a composite expression shall not be assigned
    to an object with wider essential type
  - Fixed by typecasting to the required datatype

+++ include/lib/bitops.h
@@ -93,7 +93,7 @@ static inline u32 get_mask(u32 end, u32 start);
  */
 static inline u32 get_mask(u32 end, u32 start)
 {
-	return ((1U << (end - start + 1U)) - 1U) << start;
+	return (u32) ((1U << (end - start + 1U)) - 1U) << start;
 }
 
 /**
